refactor: split BaseStaticResource helpers into traits to cut class complexity
